pyproject.toml
[project]
name = "cinepolis-reservation"
version = "0.1.0"
description = "Cinépolis reservation RESTful system."

authors = [
    {name = "Yuri Teixeira", email = "yuriteixeirac@proton.me"}
]

readme = "README.md"

requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg (>=3.3.3,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "mysqlclient (>=2.2.8,<3.0.0)",
    "redis (>=7.3.0,<8.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
]

[tool.poetry]
packages = []

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
